fix: address PR review comments for warehouse locations
- Replace selectinload with joinedload for better query efficiency
- Use Decimal instead of float for visual coordinates (x, y, width, height, rotation)
- Use commons WarehouseStructureCode enum instead of duplicate GQL enum
- Change quantity from int to Decimal for fractional quantities
- Remove duplicate WarehouseStructureCodeGQL and WarehouseMemberRoleGQL enums
- Add migration for quantity column type change (INTEGER to NUMERIC)

+++ app/graphql/v2/core/warehouses/mutations/warehouse_location_mutations.py
@@ -1,5 +1,6 @@
 """GraphQL mutations for warehouse locations."""
 
+from decimal import Decimal
 from uuid import UUID
 
 import strawberry
@@ -80,7 +81,7 @@ class WarehouseLocationMutations:
         self,
         location_id: UUID,
         product_id: UUID,
-        quantity: int,
+        quantity: Decimal,
         service: Injected[WarehouseLocationService],
     ) -> LocationProductAssignmentResponse:
         """Assign a product to a location."""
@@ -93,7 +94,7 @@ class WarehouseLocationMutations:
         self,
         location_id: UUID,
         product_id: UUID,
-        quantity: int,
+        quantity: Decimal,
         service: Injected[WarehouseLocationService],
     ) -> LocationProductAssignmentResponse:
         """Update product quantity at a location."""
